⬆️ NPM: Update @tinacms/cli to 1.9.3 and tinacms to 2.7.3 (#3759)
Fixed #3760 

This pull request includes several updates to dependencies. The most
important changes include updating the versions of `tinacms` and related
packages, as well as the `@tinacms/cli` package to resolve issues with
editors opening TinaCMS and not being redirected to the default branch
if the branch they last edited no longer exists.

Dependency updates:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L33-R33):
Updated `@tinacms/cli` from `1.9.0` to `1.9.3` and `tinacms` from
`2.6.1` to `2.7.3`.
[[1]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L33-R33)
[[2]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L112-R112)

Extension configuration:

*
[`.vscode/extensions.json`](diffhunk://#diff-c16655a98a3ee89a7636a59c59a72b0e93649e3a1e947327cfc43a1336b4e912L7-R7):
Removed the `arcanis.vscode-zipfs` extension as the project does not use
Yarn