Fixing the issue with 500 error on unknown routes (#3754)
- Affected routes: `/*`
- Fixed #3751 

The 500 error is occurring due to Next.js's `dynamicParam`. It attempts
to generate a page at runtime, but Tina is missing the relevant page
data, which is causing the issue.



✅ Added a service to retrieve Tina data and route to 404 if it doesn't
find it.
✅ Removed `wrapFieldsWithMeta`, which was causing an additional 1MB+ of
unused JavaScript across multiple routes. This change reduced the unused
bundle size from ⬆️4.77MB to ⬇️2.11MB, saving approximately 🚀 2.5MB in
total across multiple pages.

---------

Co-authored-by: Caleb Williams [SSW] <65635198+Calinator444@users.noreply.github.com>