fixed z-stacking for bleed images (#3763)
<!-- describe the change, why is it needed and what does it accomplish
-->
<!-- As per rule https://www.ssw.com.au/rules/over-the-shoulder-prs -->
<!-- Getting the PR merged is part of the PBI - Call someone to review
your changes to get them merged ASAP -->


### Description
The bleed effect wasn't working because the z-index of the image was
lower than the z-index of the background below it. To fix this I've
added a z-index of 0 to the root element to establish a stacking
context, added a z-index of 20 to the bleeding image, and added a
z-index of 30 to the wrapper for the consulting page components to
ensure they always sit on top of the image.

- Affected routes: <!-- E.g. `/offices/brisbane`  -->

- Fixed #3637


- [x] Include done video or screenshots


![image](https://github.com/user-attachments/assets/5117d00a-c1bd-46f3-abc2-ef0d770eda77)
**Figure**: **❌Before**


![image](https://github.com/user-attachments/assets/02850227-3aa0-4671-906e-668f9b847b8d)
**Figure**: **✅After**