fixed consulting page button outline (#4018)
<!-- describe the change, why is it needed and what does it accomplish
-->
<!-- As per rule https://www.ssw.com.au/rules/over-the-shoulder-prs -->
<!-- Getting the PR merged is part of the PBI - Call someone to review
your changes to get them merged ASAP -->
### Description

I noticed the outline effect on some of the consulting page buttons had
gone missing. This seems to be an issue with `tailwind-merge` not
recognising `outline` as a legitimate class. Guessing this will be fixed
when we upgrade to tailwind 4 on the website.

- Affected routes: /consulting/net-upgrade

- Fixed #3756


### Screenshot


![image](https://github.com/user-attachments/assets/1180f9d4-21cd-4698-bca0-54f978b33af3)